+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All other sections are for end-users.
 
 ### Added
 
+- Implicit `ptime` column (processing time) for streams without `ROWTIME` keyword (event time) ([#195](https://github.com/SpringQL/SpringQL/pull/195))
 - `BLOB` type ([#187](https://github.com/SpringQL/SpringQL/pull/187))
 
 ### For developers